Use Trait to add product conversion to document logic
